PR Corrections V3: extracted some PropTypes logic to types folder for reusability
