+++ react/game/src/app/components/CustomFormFields/Input/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import inputPropTypes from '@types/Input';
 
 import styles from '../styles.scss';
 
@@ -26,21 +27,8 @@ function Input({ label, input, type, meta, placeholder, id }) {
 }
 
 Input.propTypes = {
-  label: PropTypes.string.isRequired,
-  input: PropTypes.shape({
-    name: PropTypes.string.isRequired,
-    onBlur: PropTypes.func,
-    onChange: PropTypes.func,
-    onDragStart: PropTypes.func,
-    onDrop: PropTypes.func,
-    onFocus: PropTypes.func,
-    value: PropTypes.string
-  }).isRequired,
-  type: PropTypes.string.isRequired,
-  meta: PropTypes.shape({ error: PropTypes.string, active: PropTypes.bool, touched: PropTypes.bool })
-    .isRequired,
-  placeholder: PropTypes.string,
-  id: PropTypes.string
+  ...inputPropTypes,
+  type: PropTypes.string.isRequired
 };
 
 export default Input;
