+++ react/game/src/app/components/CustomFormFields/Textarea/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import inputPropTypes from '@types/Input';
 
 import styles from '../styles.scss';
 
@@ -28,22 +29,9 @@ function Textarea({ label, input, meta, rows, resize, cols, placeholder, id }) {
 }
 
 Textarea.propTypes = {
-  label: PropTypes.string.isRequired,
-  input: PropTypes.shape({
-    name: PropTypes.string,
-    onBlur: PropTypes.func,
-    onChange: PropTypes.func,
-    onDragStart: PropTypes.func,
-    onDrop: PropTypes.func,
-    onFocus: PropTypes.func,
-    value: PropTypes.string
-  }).isRequired,
+  ...inputPropTypes,
   rows: PropTypes.number,
-  cols: PropTypes.number,
-  meta: PropTypes.shape({ error: PropTypes.string, active: PropTypes.bool, touched: PropTypes.bool })
-    .isRequired,
-  placeholder: PropTypes.string,
-  id: PropTypes.string
+  cols: PropTypes.number
 };
 
 export default Textarea;
