+++ react/game/src/app/screens/Profile/components/ProfileEdit/components/ProfileEditForm/layout.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Field, reduxForm } from 'redux-form';
+import { reduxForm } from 'redux-form';
 import formNames from '@constants/formNames';
-import Button from '@components/Form/Button';
+import Button from '@components/CustomFormFields/Button';
 
+import { mapField } from './utils';
+import { formFields } from './constants';
 import styles from './styles.scss';
 
 function ProfileEditFormLayout({
@@ -12,25 +14,13 @@ function ProfileEditFormLayout({
   submitting,
   errorMessage,
   successMessage,
-  activateEditProfile,
-  formFields
+  activateEditProfile
 }) {
   const { btnTooltips, btnTitles } = formNames.EDIT_PROFILE;
   return (
     <div className={styles.editProfileFormContainer}>
       <form className={styles.editProfileForm} onSubmit={handleSubmit}>
-        {formFields.map(field => (
-          <Field
-            key={field.id}
-            type={field.type}
-            validate={field.validate}
-            component={field.component}
-            name={field.name}
-            id={field.id}
-            label={field.label}
-            placeholder={field.placeholder}
-          />
-        ))}
+        {formFields.map(mapField)}
         {successMessage && <div className={styles.successMessage}>{successMessage}</div>}
         {errorMessage && <div className={styles.errorMessage}>Error: {errorMessage}</div>}
         <div className={styles.editProfileBtnContainer}>
@@ -64,27 +54,7 @@ ProfileEditFormLayout.propTypes = {
   submitting: PropTypes.bool.isRequired,
   errorMessage: PropTypes.string,
   successMessage: PropTypes.string,
-  activateEditProfile: PropTypes.func.isRequired,
-  initialValues: PropTypes.shape({
-    firstName: PropTypes.string,
-    id: PropTypes.number,
-    lastName: PropTypes.string,
-    age: PropTypes.string,
-    aboutMe: PropTypes.string,
-    profilePicture: PropTypes.string,
-    backgroundPicture: PropTypes.string
-  }),
-  formFields: PropTypes.arrayOf(
-    PropTypes.shape({
-      type: PropTypes.string.isRequired,
-      validate: PropTypes.array,
-      component: PropTypes.func.isRequired,
-      name: PropTypes.string.isRequired,
-      id: PropTypes.string,
-      label: PropTypes.string,
-      placeholder: PropTypes.string
-    })
-  )
+  activateEditProfile: PropTypes.func.isRequired
 };
 
 export default reduxForm({
